Add files via upload
The Digital Imaging and Communications in Medicine (DICOM) standard is an image format used in medical imaging to store data. Sometimes, it is necessary to use commercial tools to view or analyze DICOM files. However, thanks to Python and its library PyDicom, you can view and analyze these files for free. A Jupyter Notebook, especially when using the Colab platform, is ideal for visualization and analysis tasks.